Big change - all client files converted to c++
This required modifications due to tighter type constraints

As a bonus, there is lower dependence on SDL types throughout the codebase.

+++ Client/optimizer.cpp
@@ -6,25 +6,25 @@ struct tri_data
 {
 	bool added;
 	float score;
-	Uint32 verts[3];
+	uint32_t verts[3];
 };
 
 struct vert_data
 {
 	float score;
-	std::set<Uint32> remaining_tris;
+	std::set<uint32_t> remaining_tris;
 };
 
 #ifdef	USE_BOOST
-float calculate_average_cache_miss_ratio(const boost::shared_array<Uint32> &indices,
-	const Uint32 offset, const Uint32 count, const Uint32 cache_size)
+float calculate_average_cache_miss_ratio(const boost::shared_array<uint32_t> &indices,
+	const uint32_t offset, const uint32_t count, const uint32_t cache_size)
 #else	/* USE_BOOST */
-float calculate_average_cache_miss_ratio(const Uint32* indices, const Uint32 offset,
-	const Uint32 count, const Uint32 cache_size)
+float calculate_average_cache_miss_ratio(const uint32_t* indices, const uint32_t offset,
+	const uint32_t count, const uint32_t cache_size)
 #endif	/* USE_BOOST */
 {
-	std::vector<Uint32> cache(cache_size, 0xFFFFFFFF);
-	Uint32 i, j, cache_ptr, cache_misses;
+	std::vector<uint32_t> cache(cache_size, 0xFFFFFFFF);
+	uint32_t i, j, cache_ptr, cache_misses;
 	bool cache_hit;
 
 	if (count <= cache_size)
@@ -58,37 +58,37 @@ float calculate_average_cache_miss_ratio(const Uint32* indices, const Uint32 off
 static const float score_table[33] =
 {
 	0.0f,
-	2.0f / sqrt(1.0f), 2.0f / sqrt(2.0f), 2.0f / sqrt(3.0f), 2.0f / sqrt(4.0f),
-	2.0f / sqrt(5.0f), 2.0f / sqrt(6.0f), 2.0f / sqrt(7.0f), 2.0f / sqrt(8.0f),
-	2.0f / sqrt(9.0f), 2.0f / sqrt(10.0f), 2.0f / sqrt(11.0f), 2.0f / sqrt(12.0f),
-	2.0f / sqrt(13.0f), 2.0f / sqrt(14.0f), 2.0f / sqrt(15.0f), 2.0f / sqrt(16.0f),
-	2.0f / sqrt(17.0f), 2.0f / sqrt(18.0f), 2.0f / sqrt(19.0f), 2.0f / sqrt(20.0f),
-	2.0f / sqrt(21.0f), 2.0f / sqrt(22.0f), 2.0f / sqrt(23.0f), 2.0f / sqrt(24.0f),
-	2.0f / sqrt(25.0f), 2.0f / sqrt(26.0f), 2.0f / sqrt(27.0f), 2.0f / sqrt(28.0f),
-	2.0f / sqrt(29.0f), 2.0f / sqrt(30.0f), 2.0f / sqrt(31.0f), 2.0f / sqrt(32.0f)
+    2.0f / sqrtf(1.0f), 2.0f / sqrtf(2.0f), 2.0f / sqrtf(3.0f), 2.0f / sqrtf(4.0f),
+    2.0f / sqrtf(5.0f), 2.0f / sqrtf(6.0f), 2.0f / sqrtf(7.0f), 2.0f / sqrtf(8.0f),
+    2.0f / sqrtf(9.0f), 2.0f / sqrtf(10.0f), 2.0f / sqrtf(11.0f), 2.0f / sqrtf(12.0f),
+    2.0f / sqrtf(13.0f), 2.0f / sqrtf(14.0f), 2.0f / sqrtf(15.0f), 2.0f / sqrtf(16.0f),
+    2.0f / sqrtf(17.0f), 2.0f / sqrtf(18.0f), 2.0f / sqrtf(19.0f), 2.0f / sqrtf(20.0f),
+    2.0f / sqrtf(21.0f), 2.0f / sqrtf(22.0f), 2.0f / sqrtf(23.0f), 2.0f / sqrtf(24.0f),
+    2.0f / sqrtf(25.0f), 2.0f / sqrtf(26.0f), 2.0f / sqrtf(27.0f), 2.0f / sqrtf(28.0f),
+    2.0f / sqrtf(29.0f), 2.0f / sqrtf(30.0f), 2.0f / sqrtf(31.0f), 2.0f / sqrtf(32.0f)
 };
 
-static inline float calc_new_score(const Uint32 count)
+static inline float calc_new_score(const uint32_t count)
 {
 	return 12.0f;//score_table[std::min(count, 32u)];
 }
 
 #ifdef	USE_BOOST
-bool optimize_vertex_cache_order(boost::shared_array<Uint32> &tri_indices, const Uint32 offset,
-	const Uint32 count, const Uint32 cache_size)
+bool optimize_vertex_cache_order(boost::shared_array<uint32_t> &tri_indices, const uint32_t offset,
+	const uint32_t count, const uint32_t cache_size)
 #else	/* USE_BOOST */
-bool optimize_vertex_cache_order(Uint32* tri_indices, const Uint32 offset,
-	const Uint32 count, const Uint32 cache_size)
+bool optimize_vertex_cache_order(uint32_t* tri_indices, const uint32_t offset,
+	const uint32_t count, const uint32_t cache_size)
 #endif	/* USE_BOOST */
 {
 	std::vector<float> cache_score(cache_size + 3, 0.75);
-	std::vector<Uint32> cache_idx(cache_size + 3, 0xFFFFFFFF);
-	std::vector<Uint32> grow_cache_idx(cache_size + 3, 0xFFFFFFFF);
-	std::set<Uint32>::iterator it;
-	Uint32 i, j, index;
-	Uint32 num_triangles, num_vertices;
-	Uint32 tris_left, best_idx;
-	Uint32 a, b, c;
+	std::vector<uint32_t> cache_idx(cache_size + 3, 0xFFFFFFFF);
+	std::vector<uint32_t> grow_cache_idx(cache_size + 3, 0xFFFFFFFF);
+	std::set<uint32_t>::iterator it;
+	uint32_t i, j, index;
+	uint32_t num_triangles, num_vertices;
+	uint32_t tris_left, best_idx;
+	uint32_t a, b, c;
 	float best_score, old_score, new_score;
 
 	if ((count < 3) || (count % 3 != 0) || (cache_size < 4))
